LiSBtS: Add an auto-convergence HUD
This shader is based off my old constant buffer debug shader, but is
intended for production use. It will display the current
auto-convergence pop-out value for a couple of seconds after the user
presses the convergence keys to adjust it. This allows the user to copy
down the value into the d3dx.ini if they wish to make their adjustments
permanent.