LiSBtS: Fix auto-convergence going wacko if stereo is disabled
When launching the game, after toggling full screen, etc. the
convergence would be set to the maximum 20. This was because the
code that looked for the convergence being changed by the user got
confused by it going to 0 and set a ridiculous popout bias.

Check separation is non-zero to confirm that stereo is enabled before
considering the convergence values.

This also highlights another subtle difference between getting the
convergence via IniParams and StereoParams in 3DMigoto - when stereo is
disabled StereoParams will be set to 0, but getting it via IniParams
will get the value that nvapi reports, which will be the current
convergence.