LiSBtS: Add hotkey to toggle auto-convergence
Pressing ~ will toggle auto-convergence on and off. This also adds some
state to track when it has just been toggled to display the HUD for a
few seconds showing the current state. The saved state is initialised to
true so that the HUD will not display on launch if auto-convergence is
enabled, but will display on launch if it is disabled in the d3dx.ini