LiSBtS: Add transparent black background to HUD
The HUD text can be difficult to read depending on what is behind it, so
cap the alpha value in the pixel shader to add a semi-transparent black
background around the text.

This also fixes the dimensions of each character emitted by the geometry
shader to match the character width so that the background will be one
consistent shade and won't stick out too far to the right. This also now
emits geometry for space characters so that they are shaded as well.