CB Debug Shader: Bring in changes to character geometry from LiSBtS
This brings in the change to use the correct geometry for characters
emitted by this shader, so that if someone chooses to shade the
background it will look right.

It does not bring in the change to emit geometry for a space character -
doing so would reduce the number of characters this shader can emit, and
I seem to recall it was already close to the limit.