LiSBtS: Fix unusual Unity lighting pattern (INV_PROJECTION_IN_PS)
This shader is seen in the lighting inside a mirror you can smash in the
junkyard. It is unusual because the pixel shader is performing an
inverse projection and is not using _ZBufferParams to get the linear
depth. To fix this we can also use the inverse projection matrix to get
the linear depth, but this seems more finicky than usual - the X & Y
coordinates were also necessary to pass through the inverse projection
matrix to get the correct depth, and correcting the depth after the
original inverse projection multiply didn't work as expected (or maybe I
just messed something up).

In addition to the new pattern in the shader, the depth buffer used for
the lighting inside the mirror was mono, as was the buffer that receives
the lighting from this shader - both of those are forced to stereo.