Unity5: Add unusual lighting shader from LiSBtS to template
While rare, I don't think this shader is specific to LiSBtS, but the
scripts don't support this pattern yet. Add the fixed shader to the
template in case it is used in other games. Since we were able to fix it
using the inverse projection matrix inside the shader it doesn't
inherently require any ini section, though it is possible that other
TextureOverrides may be required to force the depth and lighting buffers
to stereo.