LiSBtS: Disable auto-convergence when no Z-buffer
If the auto-convergence shader doesn't receive a Z-Buffer it will set
the convergence to the hard maximum limit, which is unlikely to be what
we want. Without a Z-Buffer the auto-convergence can't really do
anything sensible, so have the shader detect this and bail.

Make the HUD notify the user that there is no Z-Buffer if the user
attempts to modify the popout value via the convergence hotkeys, and add
an option to force this message to display any time this happens (mostly
for debugging).