LiSBtS / Auto-Convergence: Use depth buffer from both eyes
The auto-convergence would only use the depth buffer from the right eye,
which meant in some cases an object could obscure the camera in only the
left eye, which would not trigger auto-convergence.

This makes the depth-buffer downscale operate in stereo, and uses
stereo2mono to combine the final 1 pixel downscaled depth buffers from
both eyes into a single buffer that the auto convergence shader can read
to calculate an auto-convergence that should deal with objects obscuring
the camera in either eye.

This includes some detection for cases where stereo2mono may not be
working (which is known to occur on SLI without StereoFlagsDX10 =
0x00000008 and/or a CM profile... though for some reason it's working
fine on LiSBtS without either of those) and ignore the left eye, but
this part is not tested.