float_to_hex: Add tests for timer accuracy, and verbose mode
3DMigoto now was an option to write a time value to an ini param. It
stores it in seconds since launch in a 32bit floating point number to
make it easy to work with in shaders, but doing so means that we need to
consider floating point precision issues - since these are not always
intuitive, this adds a test case to explore at what point we will begin
losing precision to see if it is something we need to concern ourselves
with or not.

Until we hit 16384 (1.0 * 2^14) seconds since launch we have a
resolution of 1ms or better, but at that point the resolution will drop
to approximately 2ms, and every time we double the number of seconds
from that point on we approximately halve the precision. Since
GetTickCount() only has a resolution of 10-16ms this issue won't become
significant until 72 hours when the resolution will have dropped to
approximately 32ms (we might start noticing aliasing artefacts appearing
after 36 hours when we are on par with the system timer).

Interestingly, it seems if we stored the value in milliseconds instead
of seconds, we would last slightly longer before losing precision, but
not by any order of magnitude so this is not worth considering as a
solution - just an interesting artefact to take note of. This can be
illustrated by the fact we will have dropped to half a second of
precision after 48.5 days, but it takes GetTickCount() 49.7 days to wrap
back to zero.

Also add a verbose mode to the tool to break down the floats to show the
sign, unbiased exponent and mantissa.