MEA: Almost completely fix tile lighting
There are two parts to this fix. Firstly, the buffer in u2 in these
shaders is broken due to a driver bug that only updates the value in the
right eye when the game calls UpdateSubresource(). We fix these by
having 3DMigoto create a mono copy of them, followed by a stereo copy.
This creates a new buffer with the same value in both eyes for the
shader to use.

The second part of this commit applies a world-space adjustment to the
light positions as the shaders calculate which tiles cover them. This is
almost perfect, but there are still some tiles visible at vertical
camera angles. Making this commit now so I don't lose it.