MEA: Correct tile lighting fix
The fix from the previous commit was still slightly out - since it was
adjusting the light positions instead of the tile positions when the
camera was near a light the tiles would not be adjusted far enough, and
the tile seam could re-appear.

This is the correct fix - adjusting the tile position based on the
minimum and maximum depths (found in the t0 texture's red and green
channels). To get the linear depth for the stereo formula we convert the
clip space coordinate to world space, normalise it and then convert it
back to clip space.

The real trick is spotting the rather unusual matrix multiply in these
shaders, since it is performed across multiple compute shader threads
(an optimisation that we've defeated, but unlikely to make any
noticeable fps difference).