Add StereoFlagsDX10 test shaders
This adds a set of debug shaders and d3dx.ini configuration that we can
use to test the results that different combinations of StereoFlagsDX10
on buffers and structured buffers.

It creates various buffers and structured buffers with varying surface
creation mode flags and bind flags, writes to them from pixel or compute
shaders with which eye was running then displays the results. This
allows us to clearly see whether the buffer was stereo or mono.

Some interesting results:
StereoFlagsDX10 = 0x00004000: All UAVs are stereoised, all compute
shaders are run in both eyes

StereoFlagsDX10 = 0x0000C000: UAVs are only stereoised if they have the
RTV bind flag. Notably, this bind flag CANNOT be used with
StructuredBuffers, so if you need a stereo StructuredBuffer, do not use
this combination.

StereoFlagsDX10 = 0x00000000: UAVs are only stereoised if the surface
creation mode forced them to stereo, and only if they were regular
buffers, not structured buffers. Compute shaders only run for the right
eye.

The stereo creation mode can force any buffer to mono, but can only
force buffers to stereo if they had the RTV bind flag, which of course
is not possible for structured buffers.