extract_unity55_shaders: Add support for Unity 2017.1.0
This adds support for the changes 2017.1.0 & 2017.1.1 brought to the
shader file format. In particular, samplers now have a separate bind
info section from textures, and there is an extra align performed at one
point.

There is nothing inside the shader format to indicate which version it
is - I believe this is intended to come from the UUIDs in the type table
inside the main asset file it was extracted from (especially as asset
bundles appear to have a self-describing type table), so to pass that
from the first script to the second, it is now written in the filename
of the shaders.