extract_unity53_shaders: Handle DX9 shaders with debug info
Since the Unity 5.3+ format is binary, we want an assert to make sure we
have actually got to a shader binary when we think we have. For DX11
shaders we can just look for the DXBC signature, but not so for DX9 - so
we look for the CTAB section a little way into the shader instead.

One of the DX9 Hidden/GlobalFog shaders in Salt included debug info,
which was included in the file before the CTAB section, causing the
assert to fail. I still don't want to remove the assert unless I have
to, so for now add "DBUG" as an alternate section it can match.