unity_asset_bundle_extractor: Add --dump-type-info option
This option dumps the type info embedded inside Unity Asset Bundles to
disk, in the hopes that we can decipher the format and use it as a way
to future-proof the shader extraction. This is saved externally, since
it is not present in regular asset files, so we will need to reuse it
from any asset bundles we find.

Additionally, a --decode-type-info option is added to decode a
previously dumped file.