unity_asset_extractor: Don't abort if type info parsing fails
We don't handle all the variations for non-shader type info yet and our
asserts can trigger. I want the asserts left in, because we need to
catch any variations in the format and work out how to handle them, but
we also shouldn't have it stop the extraction of other types, so just
print out the error and continue.