extract_unity53_shaders: Add support for Unity 2017.3
Unity 2017.3 adds some new fields to the 5.3 style header section that
this consumes. It tags SetTexture lines in the 5.3 header section with
MSAA for those that use it, but for now just bails if any of the new
fields are set in the constant buffers - they are likely related to the
new structure definitions, but I'll need to see one in use to decode it
properly.