extract_unity55_shaders: Cope with shaders that have an asset name
Starting with Unity 5.5 shaders stopped storing an asset name in the
usual location, and instead we find the shader name much further down
into the file. But the asset name field for this was left in, and it
turns out there are still some shaders in the wild that use it, so we
must still parse it, even if we no longer use it.