extract_unity53_shaders: Add support for sampler bind info
Unity 2017.3's Hidden_PostProcessing_TemporalAntialiasing has a sampler
bind info in the 5.3 style headers that we need to parse. I'm not
exactly certain of the meaning, since the texture bind info already
contains sampler bind info, and this one was binding sampler "85" to
slot 0 via this section, whatever that means.

Annoyingly, this makes the bind info sections variable length.