extract_unity53/55_shaders: Add/fix support for structs in 2017.3
CosmicTrip's WaterOasis shader uses the new structs in Unity 2017.3 that
allowed me to decipher the right format.

Fixes crash and wrong indentation parsing these in extract_unity55_shaders

Adds parsing the 5.3 style headers to extract_unity53_shaders.

If a constant buffer contains a struct it will be displayed like:

    ConstBuffer "foo" size
    ... vector / matrix entries not in struct ...
    Struct offset [bar] array_len struct_size
    ... vector / matrix entries inside struct ...

The only example I've seen had the struct starting at offset 0, so I
can't be certain what it would look like if it wasn't, but I assume the
entries inside the struct will be relative to the start of the struct,
not the whole cb.

This means the last new field in 2017.3 that we haven't identified yet
is before the constant buffers, not after them.