Game List: Cope with blogger.blogs.get API returning errors
We use this API to retrieve the total number of posts to display a
meaningful progress indicator, however as of February 2nd 2018 this API
has begun returning 500 errors for no apparent reason. Unlike the
previous issues we've had with the Blogger API, this doesn't seem
related to any sort of timeout - Google has simply broken this API.

We don't strictly need the total posts to display the game list, so if
an error occurs just display the progress as "0/???" instead and
continue.