Custom Shaders: Add shader to move objects from keyboard input
This shader drives a transformation matrix, that can be applied to an
object in local, world or view space to move it around the world -
useful to bring broken objects closer to the camera for inspection, or
just for fun. But beware the caveat applies that by moving an object,
you may just break the effect more.

In the shader declare:

    struct transformation_buffer {
    	matrix transformation;
    };

    StructuredBuffer<transformation_buffer> transform : register(t100);

Then apply the transformation matrix somewhere of interest, e.g.

    // Using bone matrices. Posed, but broken:
    r0.xyz = mul(float4(r0.xyz, 1), transform[0].transformation);
    // Skipping bone matrices. Neutral pose, but not broken::
    //r0.xyz = mul(v0.xyzw, transform[0].transformation);

You can't apply this to an output position directly. If you wanted to do
that you would have to project it, something like:

    o5 += mul(mul(float4(0, 0, 0, 1), transform[0].transformation), g_vsScene.m_projection);