MGSV: Partially fix soldier outlines (game bug)
The latest 3DMigoto provides the tools necessary to get more data on
this issue. The CB debug shader shows that the bone matrix for bone 0
has a large transformation applied that doesn't make sense (40+ units
towards the camera), though is notably similar to the transformation of
the entire soldier in at least one case I checked.

This partially fixes the issue by using the bone 1 matrix instead of
the bone 0 matrix. This will result in a partially broken outline, but
this is unlikely to be noticeable unless someone is looking for it, and
much better than the original broken effect.

This also includes debug shaders that can freely move these outlines
around the world via key presses, so that they can be brought closer to
the camera for examination, and rotated to see the distortion from other
angles.