Subnautica: Fix Screen Space Reflections (CB steroisation technique)
This brings a variation of my UE4 constant buffer stereoisation
technique to Unity. It uses a compute shader to generate stereoised
matrices for some of the matrices found in UnityPerDraw and
UnityPerFrame, that when applied to the Screen Space Reflections will
accurately correct them.

This may work with other Unity shaders as well as it does with most UE4
shaders, but the fps cost associated with redirecting them to texture
buffers may be too high (this is already costing several fps on the
menu) - this can be reduced by restricting the redirects to only the
matrices of interest.