asmtool: Allow cb remaps to use different offsets in shader resources
We might want to remap a matrix from somewhere in a constant buffer to a
different location in the shader resource buffer, so add an extra
parameter to the remap commands to allow this. Do this now before they
are used so we can always require this parameter, even if they are not
being moved.