Subnautica: Have CB stereoise shader cope with no projection matrix
This was a bit of an experiment to see what challenges we would face if
we applied the CB stereoisation technique to Unity, and as predicted the
big challenge is that Unity only updates the matrices when a shader
actually uses them, unlike UE4 where we know the view buffer is valid
through the whole frame.

The water surface SSR shader used on the main menu uses the projection
matrix, so it is valid, but in game SSR is not used and the projection
matrix is stale.

Instead, we derive it ourselves. This isn't as simple as doing
inverse(view) * view-projection unfortunately - it seems we need a
cleaned up projection matrix. That does get us the FOV though, so we
still do it for that, but then we create a new projection matrix using
those FOV values and the near and far clipping planes derived by
normalising Z values through the view-projection matrix and its inverse.