Subnautica: Fix underwater sun shafts (CB stereoisation technique)
This effect uses some custom CameraToXXX matrices in the Globals CB that
need to be stereoised to pop the rays into 3D. This follows a similar
concept to the last CB stereoisation shader, but instead aims to create
an inverse-stereo-view-injection matrix.

The biggest gotcha here (that was not entirely unexpected from what I
know of Unity) is the view-projection matrix unity provided in their
"per frame" constant buffer is not valid despite being used by the
shader - it is a full screen matrix instead.

In this case it seems ok if we don't bother creating a cleaned up
projection matrix - it only slightly affects the ray position, and not
enough to justify the extra maths.

Although an interesting experiment, I cannot really recommend that other
shaderhackers apply this technique manually like this - there is a cost
in fps, and doing this separately for every broken effect may not be any
easier than traditional fixing techniques - I recommend restricting this
to the more difficult effects to fix in the traditional manner, such as
screen space reflections. This experiment was in part to test out the
viability of possible future modding tools based around the concept.