Subnautica: Adjust mouse depth while in tablet
While the tablet is displayed, we want to push the mouse cursor to
roughly the same depth, so enable the software mouse and add the
appropriate adjustment in the tablet preset.

The custom mouse cursor this game is using has garbage in the
transparent pixels and lacks a mask, so we need to adjust mouseps.hlsl
and the blend state to assume that it has an alpha channel, and is not
pre-multiplied.