Subnautica: Swap inventory icon exclusion to Y-cutoff instead of texture+shader filter
The battery swap HUD uses the inventory icon shader, and since it has
surrounding text that is adjusted we need to adjust the battery icons as
well, so enable the adjustment for this shader.

Switch to using a Y cutoff to prevent adjusting the inventory icons at
the bottom of the screen.