Subnautica: Capture depth of in-world menus to consider for HUD depth
There's a couple of in-world menus in the game (e.g. upgrade station in
the moonpool, and some of the controls in the Cyclops cockpit) that we
want to consider when deciding on the HUD depth.

This runs a second pass for each HUD shader with a custom depth buffer
bound to capture the depth of these HUD elements. The hud.hlsl checks if
we are running the second pass and only allows elements that are drawn
in-world to be rendered to this depth buffer. No pixel shader or render
target is bound to keep the cost of this pass low.

To allow hud.hlsl to distinguish between the two passes we need to set a
flag, however since there are a lot of HUD elements using IniParams for
this added up to a measurable few fps. Instead, we simply check if the
HUD depth buffer is bound to the t111 slot or not, since that doesn't
require any additional CPU to GPU transfers beyond the commands to bind
and unbind it. It's worth noting that a similar technique can be used by
creating a custom resource that has the sole purpose of being tested
from the shader and binding and unbinding it from 3DMigoto instead of
setting an IniParam.

This does not auto-adjust the HUD depth when the colour customisation
terminals are used, as there is no immediately apparent way to
distinguish between that scenario and bringing up the esc menu.

This will need testing on non-SLI, as unbound pixel shaders and render
targets may cause the driver heuristics to behave differently.