Subnautica: Disregard invisible in-world HUDs
The Cyclops cockpit has a warning display right in the middle that
is usually invisible, only turning on when a collision is immanent.
However, the game still calls the draw call for this even when it is
invisible, so we end up with it on our depth buffer, and since it is
right in the middle of the display it tends to mess up the auto HUD.

Test for this based on the alpha channel of the "colour" the game passed
in, and do not consider depth of any that are fully transparent.