Subnautica: Use HUD depth from previous frame for better performance
With the scanner HUD on the fps dipped by almost 5fps due to the HUD
depth pass. To regain this, we no longer intermix reading and writing to
the same depth buffer, as doing so presumably creates memory barriers on
the GPU and becomes a bottleneck. Instead, we write the depth
information to one buffer and read from a second buffer created in the
previous frame - that minimises the memory barriers and regains the lost
performance.