Subnautica: Enable crosshair by default
There are some in-world control panels (door keypads, Cyclops HUD, etc)
that are easier to manipulate with the crosshair visible. At other times
the crosshair is still annoying, so we want the toggle, but switch to
having it enabled by default so users know it's there and have to press
X to disable it, rather than having it disabled and having someone who
didn't read the README have a hard time.