3DMigoto-stats: Remove alpha channel on produced graphs for geforce forums
gitstats / gnuplot is producing graphs with a transparent background,
which is fine if they are displayed on a white web page, but doesn't
work on the geforce forums. It sounds as though it is hit or miss as to
whether gnuplot can even set the background of a png image any more, so
use imagemagick to remove the alpha channel after producing the images.