shadertool: Fix incorrect mask used in --fix-unreal-halo-sct, crashing Helix Mod
This would occasionally adjust an input using a bad mask copied from a
swizzle used elsewhere in the shader:

    mov r30.xxyw, v4.xxyw

This would lead to the shader failing to assemble, and if it was in
ShaderOverride that would lead to Helix Mod crashing on F10 reload.

Look up the dcl_texcoord and use the mask from there instead.