DOAXVV: Increase HUD cutoff depth to catch confetti
The confetti on the victory screens is rendered very close to the
camera, using the same shader as for the mouse sparkle and other
effects.

It is 3D, but too close to easily see, and since it is right on the
cutoff depth, some confetti pieces get stretched. Increase the HUD
cutoff to catch all the confetti and return it to screen depth.