DOAXVV: Change to new mechanism to force full screen from upcoming 3DMigoto
full_screen=1 interferes with the launcher (buttons don't work, and have
to alt+tab out to return it to windowed mode), and doesn't work after
the main game has been alt+tabbed out.

This uses a new feature of the upcoming 3DMigoto 1.3.5 to force full
screen on a key press, so once in game pressing F7 will instantly switch
to exclusive full screen, and this can be repeated after an alt+tab as
needed.