DOAXVV / Auto-Convergence: Fix stutter during slot convergence transitions
Since the convergence is set asynchronously, after the shader sets a
convergence, the next frame not use that convergence - rather, it will
continue to use the same convergence as the previous frame. The
auto-convergence shader for that frame will then calculate a new
convergence value based on the old convergence, but will have a new
depth buffer, so it can get a different result than the previous frame,
even going backwards in many cases causing the convergence transition to
appear to stutter.

To fix this, we have the auto-convergence shader use the convergence
that it set in the previous frame instead of the current convergence.
This results in smoother convergence transitions and eliminates the
stutter.

Note that if the convergence took more than one frame to propagate back
to the CPU this measure would be defeated - to counter that we would
either have to use the last convergence value it set regardless of how
many frames ago it was, or have 3DMigoto inform the shader if the
transfer has completed.