DOAXVV: Fix crash on launch with allow_create_device=2
The launcher uses a DX10 feature level, which introduces problems like
preventing the command list from being able to create shaders and
crashing DirectX on the draw calls in the custom shader sections. These
work fine if they are only enabled after the main game has launched, as
it uses DX11 properly. This is revealed now after the addition of
auto-convergence, but the same crash occurs if any of our other present
time shaders are enabled, such as the software mouse or 3DVision2SBS.

This forces the DX11 feature level to resolve these crashes, however
note that the debug DLL still crashes with a report of a DeviceChild
refcount prematurely going to 0. This may be due to key differences in
the refcounting between DX10 and DX11, and so may need a better fix
inside 3DMigoto to cope with these.