Auto Convergence: Ignore previous frame if stereo wasn't enabled
When switching to full screen the auto convergence shader would
mistakenly think the user had adjusted the convergence, and would change
the popout bias based on that, which could result on the convergence
target being not-ideal.

To detect this, record if stereo was enabled in the previous frame
(required 3DMigoto 1.3.6) and disregard apparent user convergence
changes on the frame where it is enabled.