DOAXVV: Fix ripples
The ripples were rendering at screen depth. They are first rendered to a
separate buffer in 3D, which the water shader uses to draw them, so by
forcing this first buffer to stereo we get stereo ripples.

StereoMode=1 is not enough to force this buffer to stereo, because the
ripples are rendered with no depth buffer bound, so the driver doesn't
apply the stereo correction. We can use the dummy depth buffer method to
solve this, but since the shaders used to render these ripples are used
for a lot of things, do this in the TextureOverride for the ripples and
use checktextureoverride to run it.

It seems the game renders specific ripples at varying depths that don't
quite line up with the surface, so this isn't perfect, but at least the
main ripples are lining up properly.