DOAXVV: Alternate ripple fix
To avoid the game bug where specific ripples are rendered underneath the
surface, we revert the ripple render target back to mono, and apply a
correction in the water pixel shader to account for the 2D coordinates
instead.

This adjustment is effectively undoing the halo correction in the vertex
shader for the ripples, but using it for the refraction. Alternatively,
we could forgo the adjustment in the vertex shader and correct the
refraction in the pixel shader. The point is - this is one explanation
for why we sometimes may need adjusted and unadjusted coordinates -
ideally we should only have adjusted coordinates, but that requires that
all inputs are correctly stereo, and in this case doing so reveals a
game bug, while leaving it mono works around that.