Text Shaders: Increase per-invocation character limit to 42
The maximum number of characters that a single invocation of the text
geometry shader can produce is 36, which is dictated by the size of the
buffer passed to the pixel shader. To output more characters, the
geometry shader needs to be invoked multiple times (e.g. by increasing
the number of vertices in the draw call), however the additional
invocations of the geometry shader are independent of each other and
can't continue a line of text (at least not without re-calculating the
cursor position based on the text the previous invocations would have
emitted).

We can increase this limit slightly by reducing the size of the gs2ps
buffer. This eliminates the "character" field passed to the pixel
shader, and moves the position calculations based on it into the
geometry shader so they are combined in to the existing texcoord output.
This gives enough room for an extra 8 characters per invocation.