Text shaders: Increase per-instance character limit from 42 to 51
This gets a little tricky to reduce the size of the buffer passed from
the geometry shader to the pixel shader to increase the character limit
per invocation of the geometry shader up to 51.

SV_Position has to be a four component floating point value - that seems
to be non-negotiable (also, reducing it to a half4 doesn't help, since
the 1024 limit is related to the total number of scalars the geometry
shader outputs, not the size of the buffer). However, we don't really
care about the Z and W components of SV_Position, which we just set to 0
and 1 respectively.

If we can use these unused components to store one or both of the
texcoords, that can eliminate the need to pass these separately,
thereby increasing the character limit, and it turns out we can pretty
easily use the Z component by disabling Z clipping. The W component is a
bit harder due to the perspective divide, but just being able to use the
Z component means our character limit is raised to 51.