Text shaders: Raise per-invocation character limit from 52 to 64
This gets even more tricky to minimise the size of the buffer passed
from the geometry shader to the pixel shader so that it is just
SV_Position (which is required if using the rasterizer, so this is the
minimum theoretically possible with this approach), allowing each
invocation to render up to 64 characters.

To achieve this, we use the SV_Position's W component as the Y texture
coordinate, however since this is used for the perspective divide we
have to be careful not to write a 0 to it, and take steps to counter
floating point precision issues. To this end we bias the texture
coordinate by adding a fixed constant to it, and subtract that back off
when retrieving the texture coordinate. This value must be large enough
so that floating point error does not make any noticeable difference to
the text, so a fairly generous 4096*font_scale is used by default.