DOAXVV: Prevent burst preset from interfering with auto-convergence popout
We cleared the w4 value in the burst preset to try to prevent the
auto-convergence shader from interpreting the convergence change as a
user convergence adjustment, but since presets restore the previous
value when they complete it would be set again, defeating that.

Now that presets can run command lists we can use that to make the
setting persistent after the preset finishes. It can't be set in the
preset, as that would still restore the previous value (which might be a
behaviour worth revising in 3DMigoto), but setting it from only the
command list works.