Blender 3DMigoto Addon: Add ability to remap vertex groups on export
This adds a new import option to the Blender addon to load a vgmap file,
which is just a json encoded file containing a mapping of vertex groups.
This will be added to a custom property of the object, and when the
object is exported it will be applied to alter the vertex groups.

This is useful for when essentially the same mesh is used in multiple
situations (e.g. different clothes) with differing vertex groups.

A suffix can be added when importing the vgmap file, which will be used
by the export process to generate multiple vertex buffer files in batch.