Blender 3DMigoto Addon: Add option to automatically load related meshes
Some games may divide up a mesh into multiple sections that are drawn
individually, that can be identified by using the same vertex and index
buffer as each other, but a different range of vertices/indices selected
in their draw calls.

Depending on why the mesh is being imported to Blender, it may be
desirable to load these related meshes together or separately. This adds
an option to automatically import other meshes sharing the same
vertex/index buffer as that selected.

This will not work well in games that store unrelated meshes in the same
buffers.