Blender 3DMigoto Addon: Option to load multiple meshes as separate objects
When we are loading multiple meshes at the same time we don't
necessarily want to merge them together as one, since if the game
separated them into different draw calls it may have had a good reason,
like that they are used with different shaders, materials, or are just
completely different objects.

This adds an option to select whether meshes loaded together are merged
into a single mesh (to stitch together meshes split across multiple draw
calls just because of some triangle limit), or loaded as separate
objects (which is the default).

Since this new behaviour makes it much safer to load other meshes from
the same buffers, the auto-load related meshes option is now enabled.