Blender 3DMigoto Addon: Fix exported vertices vs loops
Blender stores the vertex position in each vertex, but normals, tangents
and UV coordinates are stored in the vertex loops. Conversely, each
vertex we create when exporting must be a unique set of all parameters.

If we export an unmodified mesh that we imported this won't matter, but
if the mesh has been modified in certain ways, or the artist used the
"remove doubles" button then this will be a problem.

This fixes the vertex creation to potentially create multiple DirectX
vertices from each Blender vertex if the data in the loop vertices
serves to make them unique. It implements our own vertex de-duplication
and indexing scheme - there might be a (potentially much faster) helper
to do this for us, but it would take me longer to find out than write
it myself.