DOAXVV: Enable support for third party costume mods
This fix is now being used as the basis for costume and mesh
replacements in this game, so to make things easier this brings in the
latest 3DMigoto with lots of enhancements to support this, as well as
enabling the required stanzas to enable the mods.

The mods themselves are not distributed with this fix.