Blender 3DMigoto Addon: Cope with filenames lacking hash
If Frame Analysis dumps out a vertex and index buffer that was replaced
by 3DMigoto it will not have a hash in the filename and the addon will
hit an assertion failure.

Make the routine that looks for related files cope with this situation.

Note that the draw call count may not be right if these used a custom
draw call, since Frame Analysis uses the draw call count from the game.