Blender 3DMigoto Addon: Warn about files missing hash
Files missing a hash may signify that Frame Analysis dumped a custom
resource, which in some cases may be cut short and missing triangles -
or at least the .txt files may be - issue a warning about this.

We could also add an option to 3DMigoto to dump the full buffers (which
may also be useful to export complete meshes to single files), but given
that the .buf files do contain the complete buffers this seems a little
redundant. The .buf files do not contain enough information to load them
directly, but if we gave the importer the layout it could.

Also, it stands to reason that if Frame Analysis dumped a custom
resource, the user may have that resource on disk that we could load
directly.