Blender 3DMigoto Addon: Add option to parse .buf files instead of .txt files
The .buf files that 3DMigoto dumps for Frame Analysis are the complete
buffers that were bound to the pipeline, and may contain vertices and
faces that are not included in the .txt files, as those are limited to
only the vertices & indices actually drawn in a given draw call.

This can be useful if a sub-part of a mesh is missing from a dump (and
where merging all sub-parts is wanted), or for when a frame analysis
dump included an injected mesh where the DirectX index count may not be
accurate and the text files may be incomplete.

For the later case it is automatically enabled when loading a file that
does not include a hash, as this typically signifies that it is an
injected mesh.