Blender 3DMigoto Addon: Downgrade no hash warning to info
To try to make this addon a little easier to use, downgrade this warning
into an info. If something is wrong the user will be able to see, and
the notice will still be displayed, but no longer looks like something
has gone wrong just because a re-injected mesh has been dumped out and
loaded into Blender.