Blender 3DMigoto Addon: Use active object instead of selection when exporting
Blender can have multiple objects selected simultaneously, but only one
active. Use the active object when exporting rather than the selection,
as this way there is no ambiguity as to what to do when multiple objects
are selected, and it makes the workflow easier.

Also make the (last) imported object active when importing.