Blender 3DMigoto Addon: First pass of pose importing
This is a first attempt at importing a pose constant buffer into
Blender. It does indeed place the mesh into the correct pose dumped from
the game, but the bone placement doesn't seem right - it is actually
better than I was expecting - they aren't all at the origin and are
vaguely placed around the model, but they aren't lined up with the model
and which bone controls which vertex group is not very intuitive.

There are also some bones stretched out towards infinity that don't seem
to be linked to the mesh and would be nice to filter out.

None of the bones are linked, since that kind of relationship
information is not available in the constant buffer and the best we
could do automatically would be a guess.