Blender 3DMigoto Addon: Add operator to merge skeletons of related objects
In a game, a particular object or set of related objects may be divided
up into many draw calls, each of which may have a unique skeleton so
they can't simply be reassigned to use the same skeleton, but many of
the bones may be shared. An example of this may be an article of
clothing, that would contain all the bones of the character wearing the
clothing as well as any additional bones to animate the fabric
separately.

This operator will take multiple objects that were dumped in the same
pose and merge their skeletons together, such that common bones between
the objects are shared and unique bones are appended. This will also
remap the vertex groups as necessary since the bones may not have the
same numbering.